Add explicit http-parser, yaml-cpp rpath on linux
This tells the linker where to find those libs at runtime, which otherwise
is not known and leads to an error when executing tests.

It was not needed to do so for other libraries because those are managed
by libtool, which does this trick on the fly when executing tests.