poller: implement call_later()
Needed because in test/common/async.cpp we must schedule the test
some time in the future after the background thread was started.

In addition to be useful in this case, this function will be useful in
a variety of other use cases, so it make sense to introduce it here.