async: make sure at least one event is scheduled
We do not want the poller loop to exit, no matter how it is actually
implemented. That is, tor uses some periodic events and also some
libevent loops seems to use some periodic events. But, more in general,
we cannot count on that, and the simplest fix is to have our own
periodic event that fires every ten second.