test/async: rewrite without using delays
Reason why this is needed: with the "normal" poller we cannot
schedule events "now", as the code was doing. In fact, since the
event loop is not running yet, those events are ignored.

Instead, schedule the first events to happen a little later
in the future and, then, let test-complete callbacks schedule
new events until a total of 16 tests are run.