async: make sure we can exit background thread
I've added a destructor to Async that first breaks the poller loop
and then calls join() on the thread to wait for its completion.

I also needed to move evthread_use_pthreads() in src/common/poller.cpp
because we must ensure we have pthreads before creating the new event-base.

In fact, if this happens, the new event-base is notifiable from other
threads and hence it is possible to call event_base_loopbreak() to break
out of it from another thread. Instead, if we don't do that, -1 is then
returned by event_base_loopbreak() leading to MK raising exception.