poller: make sure event-loop does not exit
When we call loop(), register an always-on event that keeps the
poller loop alive even if we run out of events.

Required to make libevent loop equal to tor event loop as regards
the behavior when the loop is "idle" (actually tor event loop does
not exit because tor has its own periodic per-second event).

Could be implemented with less lines of code if the currently
unstable version of libevent was released as stable.

Based on a diff written by @AntonioLangiu for src/common/async.cpp.