Teach transport to record received/sent bytes
Closes #102.

To make sure I was not doing mistakes in this diff, I renamed the
transport's functions to send as write() and the actual function used
to implement send as do_send(). This way the compiler helped me to
find out all the places where refactoring was needed.

Besides, given Twisted, I prefer `transport.write` to `send`.