socks5: attach to already established connection
This diff could introduce memory leaks when/if Var<Transport> is captured inside of Socks5 code. We should check.

With this diff in tree, we can concentrate on refactoring HTTP code using a connect() function. Once that piece of work is done, it's then possible to remove the old connect code from Connection.

This diff also adds an example program that could be used to test transport in general and socks5 support in particular.

This diff also does s/socks5::/socks_/g because creating a socks5 subnamespace for a bunch of functions was over engineering.