test/http/request.cpp: improve tor related tests
Rather than having a disabled test, write out a flexible test
capable of dealing with both the case where tor is not running
and the case where it is running.

This could possibly be another way to deal with the issue
pointed out by @AntonioLangiu in #458: we can write test that
deal with the networking being down gracefully, rather than
skipping the test, and so we test that the library behaves
correctly also in the event of network being down.