big slow cleanup: ditch http/client
This diff is the first a series I have called `big slow cleanup`. They
will transform MK as we know it. They will basically finalize the changes
initiated when refactoring DNS at end of February.

In this first diff, I remove the HTTP client and replace all its usages
with the http::request() interface. I also refactor the tests and add XXX
comments in places where switching interfaces also has an impact on the
underlying assumptions made by the code, calling for more refactoring.

This is a good places to remind that all these changes are justified by
using an API that is less prone to errors and easier to use.