transport: remove factory returning ErrorOr<Transport>
This diff removes the factory that returns ErrorOr<Transport>, now that
all MK code has been migrated to the factory that takes a callback.

With this removal we can also take down lots of code that was a bit
unsafe because it bound callbacks to this.

After this removal, we can probably perform some more cleanups, but with
this I'd say we've nearly finished deleting old things.