connection: use a raw bufferevent
Using a raw bufferevent we can bind the moment where the Var<Transport>
that wraps the Connection dies and the moment where the bufferevent is
destroyed, thanks to Connection's destructor.

This is useful because it means that, after close(), the Connection and
its underlying bufferevent may stay alive, if someone holds an extra
reference to Var<Connection>. Hence everything is valid and the object
is in principle consistent except that the C level callbacks are now
gone and so it is not that much useful.

In other words: if the user does not call close(), the connection remains
open, which is quite what you would expect from a connection.

If the user calls close() and forgets about the connection, this is
destroyed (including the bufferevent) in the next I/O cycle, hence we
satisfy the requirement that bufferevent_free() shall not be called
from a bufferevent callback for SSL bufferevent does not like it.

If the user calls close() and keeps the connection alive, he is keeping
a reference to a Transport where no I/O is possible because the C libevent
callbacks have been reset, but still consistent.