build system: remove compile builtin feature
I will further refine this. The general idea is that you should
have dependencies installed beforehand, except of course those
dependencies that we compile unconditionally with MK.

This simplifies a lot and it is good to do that, because we are
about to add LibreSSL as a dependency, which is huge and would
have complicated even more the build system.