Add initial SSL implementation (#498)
* Add initial SSL implementation

This is very rough. In particular, we do not validate certificates, so
a man in the middle is very possible for now.

But I feel like it's important to have the basic scaffolding such
that things that should go in parallel over SSL can.

Crossing fingers as for Valgrind. I recall that used to reported
several uninitialized variables in OpenSSL.

* ssl-context is now non copyable and non movable

Spotted by @AntonioLangiu, thanks!