Humans: when they're in doubt, they're right
Yesterday on XMPP:

@hellais: I am confused by net::connect() in net/transport.cpp
and net::connect() in net/connect.cpp because they have the same
name and more or less same prototype but one is the interface
and the other one is the logic.

Today on MacOS:

clang++: I am confused by net::connect() in net/transport.cpp
and net::connect() in net/connect.cpp because they have the same
name and more or less same prototype but one is the interface
and the other one is the logic, but I am going to tell you this
in a very complicated way that you will spend lots of time
trying to really understand what I am saying!

And now /me needs a third coffee.