poller: use Var<> to decide deletor in constructor
Another diff functional to future templatization of this class. To implement
the template, we need to set the destructor function in the constructor since
we cannot templatize the whole class, because it is a derived class.

After this change, I think the next diff is probably going to be the real
templatization of the class with annexed cleanup of libs.hpp.