async: radically simplify
This is similar to work done in #543 and #390, except that here the
changes are small and incremental with respect to master.

As in #390, this diff changes behavior, because we always have a running
thread, which is compatible with a Tor based loop.

Unlike #543, here the patch only operates in src/common/async.cpp.

In the long run, I'd like to end up in a situation where we have #543,
but I recognize that this diff causes less changes in the tree.

Also, #543 is more fancy with respect to threads. Here I have stayed
dumb on purpose (see the 1.0 second delay when running new tests), not
to complicate to much the diff and hopefully expedite review.