protocol_impl.hpp: refactor test selection code
Refactor code selecting the test to run and eventually erroring out
such that it's easier to see what happens if a test id is not recognized.

Specifically, calling the callback of the stage with an error should
lead to the connection being closed by the client.

This diff is best read enabling the `-w` option to ignore space changes.