Revert "configure: add -pthread to flags only when needed (#607)" (#622)
This reverts commit 6ceb9e875741b1cd18f92cf520a6e9ab82e749d8 because it
was the wrong way to address the issue, unfortunately.

The warning was still present. Then, I further investigated and understood
that the way in which clang for OSX deals with -pthread is different from the
way in which the same flag is dealt with on Linux systems.

Given than the matter is more complex than I thought and that the warning
for the -pthread flag only provides some annoyance, I am reverting to before
pull request #607. We will perhaps give another stab at it, but for now
let's ignore the problem, because it's not so important.