Prevent excessive, O(N) stack consumption (#644)
Testing on device, I noticed that readn() calling its callback immediately
leads to excessive stack usage and, in the case of receiving Web100 variables
after a NDT test, could exhaust the stack of iOS devices.

Fix this by scheduling a callback for soon even when the buffer already
contains enough data to fullfill the readn() request. Testing on iOS devices
show that this fix is enough to avoid stack exhaustion and crash.