ooni: use the new report::Entry (#646)
As part of this pull request, I've split entry.hpp because that leads to better coverage measurements.

Not totally clear to me why inline templates in entry.hpp are not part of coverage. However, I've added a comment to address this issue.

Note that I am sure that those functions are executed. I say so because I also tried renaming them and changing other code accordingly, and still they resulted as not executed.