Teach reporter to notify on error
While there, make sure we initialize `test_start_time` and make
sure that a closed report is different from a report that was
not ever openned in its life.

While there, change timestamp() not to blindly throw a generic
exception but rather to return an ErrorOr<>.

While there, fixup the build badge for circle-ci which I finally
managed to correct setup and should not build all pushes.

- - -

Note: this pull request would increase the number of emitted
warnings; I will deal with them in another pull request, where
I will spend time to improve the BaseTest and OoniTest.