Unify NetTest and NetTestDsl classes
It sucked that we have a different class just to declare the DSL
when the two classes basically have same attributes.

Unify the two classes and allow the user to choose whether to
use the DSL run style or to have more control.

Not possible in one simple step to allow running OONI tests using
both DSL style and having more control.

But this is consistent with previous behavior, so no worries.

Instead possible to do this with NDT, as demonstrated by regress.