net_test.hpp: remove identifier()
We can now run the test in many ways and in some of them what is
actually run is not the test but a dynamically allocated copy.

As such, the notion of a unique identifier may lead to surprises
where the same test has two different identifiers; one when is being
scheduled and the other when it is running.

So, better to remove the functionality.