Pass timeout to underlying bufferevent
Run this experiment in the master branch:

```
time ./example/http/request -vvv https://cdn.kernel.org/pub/linux/kernel/v4.x/testing/linux-4.7-rc6.tar.xz > /dev/null
```

It will exit after five seconds because of a timeout error.

Repeat the experiment with this patch applied and see that the download
runs much more than five seconds, because now the timeout is correctly set
on the underlying bufferevent, meaning that now it is raised when there
are `timeout` seconds of silence in either direction of the socket.

While there, remove a lie in the documentation.