SSL: workaround for data delivered after EOF
See the test case in #711. It is my opinion that EOF is received by
libevent and passed straight to mk::net::Connection rather than queued
after data lingering in input buffer. The sequence of events that we
see, in fact, looks like: EOF followed by data.

To bring this under control, this diff introduce these changes:

1) introduce new LingeringData error context

2) modify net::Connection event handler to pass lingering data as
   error context for EofError

3) modify request_recv_response() to process lingering data before
   processing the actual EOF, if there's lingering data

4) modify net::Emitter to suppress events when the proper handler is
   not set, and instead log the suppressed events; this makes sense
   since in many places we set handlers to `nullptr` to state that we
   are not interested to events, so reacting ignoring is consistent

5) add some more debugging statements to net::Emitter