Make sure we copy on_entry()
For now this patch will do but we'd better properly use copy
construction to avoid losing pieces when we copy.

Also, it would probably make sense to second think the current
approach where we schedule a Var copy of the original object to
see whether there is a better approach for doing the same.