Rework diff to allow closing and reopening
Thinking more deeply at how this will be used in the future, I noticed that
just memoizing on the file names is bad because in a certain number of cases
we're going to automatically download new MaxMind databases and then we
would like to reopen them, even though the name did not change. Thus, add
a proxy class that allows memoized open() and close().