connect.cpp: increase default net/timeout
A timeout of at least 30 seconds for connect() should be a reasonable
timeout and it should in particular be more reasonable than five seconds
given that we have seen that inactive heroku machines, for example,
take more than five seconds to respond.

This should fix the immediate bug reported in #864. I've yet to decide
whether this would also mean we can close the bug or not.