connect.cpp: increase default net/timeout (#872)
* web_connectivity: set the accessible key

As reported by @lorenzoPrimi in #867, the accessible key was not
correctly set in the output JSON.

This simple diff should fix the issue by setting the accessible
key depending on whether there was blocking.

Closes #867.

* connect.cpp: increase default net/timeout

A timeout of at least 30 seconds for connect() should be a reasonable
timeout and it should in particular be more reasonable than five seconds
given that we have seen that inactive heroku machines, for example,
take more than five seconds to respond.

This should fix the immediate bug reported in #864. I've yet to decide
whether this would also mean we can close the bug or not.