web_connectivity: use copied caller options for control
After #889, I tested again on the device only to notice that the
test still does not work because the ssl-bundle option is not correctly
passed to the caller. Thus, prepare a better fix by which we leverage
that options we receive are copied and we only override the settings
we need to override, and the rest is unchanged.