http/request: don't redirect if *max_request <= 0
If the user has not asked to follow redirect (i.e. the max number of
redirects is zero or less than that), do not enter into the branch
where we follow the redirect even if there is a `location` header in
the received headers set.

Fix originally implemented to #811 and backported as hotfix.