tls_verify.c: add some casts to signed-char-*
We enable many warnings, one of which (`-Wpointer-sign` on clang) is
telling us that we're passing `data`, a `unsigned char *`, to functions
taking as parameters `char *`. Avoid the warning by converting to
`char *` which does seems to me a very safe convertion to implement.

This further modifies vendored code, but overall the difference with
upstream remains small and we can be happy with no warnings.