dns: enums that should not conflict with windows
In #842, I proposed of removing defines. After further discussions
others however I understood that it's better to keep them as they
give you certainty that a value is in a specific set.

So, since on Windows the build failed because there is a IN define
in some Windows system header, change the names used by defines.

While there, make code more robust and move implementation away
from the headers.