Better and more consistent exception handling
- never filter `std::bad_alloc` in cases where we catch most exceptions

- make sure we catch most exceptions in test final callback

Notes:

1. here I'm saying most and not all exceptions because in C++ one could
in theory raise e.g. an integer but the code always assumes that it will
catch `const std::exception &`.

2. Also, cases where a more narrow set of exceptions are considered
(meaning that other exceptions are not catched) are not handled by
this diff for obvious reasons.

3. There are cases where most likely `std::bad_alloc()` does not happen
but I have always added this special case when we were catching most
exception for robustness.