Revert more of 9b586f7
I suspect that catching exceptions in the final callback of a test
is the culprit of the travis failure observed in one build linked
with pull request #991:

  https://github.com/measurement-kit/measurement-kit/pull/991#issuecomment-265803168

To be honest, I cannot reproduce the problem locally even forcing
the specific neubot server used in the failing build.

But, in addition to this travis problem there is also that probably
catching all exceptions in the final callback does not make much
sense. In other cases, in fact, we're catching exceptions in bits
of code that are user defined and run in the context of the whole
test workflow, so now we're ignoring those exceptions but in the
future we could perhaps route many of them (e.g. by passing the
unhandled exception to the final callback), whereas with the final
callback with have little option but to ignore the exception.
Moreover, an exception in the final callback is bad because it is
the piece of code that defines what the program can do next, whereas
other exceptions in other callbacks do not influence the program
flow (as said, we can choose to route them to the final callback);
as such, it seems to me a bad idea to swallow exeptions in the final
callback.