Add new patch to disable getentropy() on ios
Regarding 4, this is inspired to the patch implemented by nickm for
tor to address the same issue, see:

    https://gitweb.torproject.org/tor.git/commit/?id=https://gitweb.torproject.org/tor.git/commit/?id=16fcbd21c963a9a65bf55024680c8323c8b7175d

Except that my patch is much more simple. It unconditionally removes
the macros for detecting getentropy. Also, we don't bother with
disabling clock_gettime for ios because I've seen that such function
is not used by the code used to emulate getentropy on ios.