getaddrinfo: android dislikes AI_V4MAPPED|AI_ALL
Honestly, those are not super important flags anyway, because they
instruct getaddrinfo to return ipv4-mapped addresses only when there
are no IPv6 addresses and one requests for AAAA.

We always request for both A and AAAA when we resolve for connecting (and
for more precise results one should probably use c-ares in future).

Sure, I could keep the feature, but then I wonder whether this would lead
to the library behavior being inconsistent across different platforms in
a not-super-simple-to-debug way.

For this reason, I'm removing the feature entirely rather than just
making it conditional on !__ANDROID__.