Fix API and ABI compatibility for old Androids (#1047)
* Improve Android build scripts

* Remove the --force flag

* Stop logging NDK dir

* No need to install libc++

* make-toolchain: the --stl is not optional

So better to put there libc++ rather than the default, which is
not the libc++ library being used to compile MK.

Doing that even though this script is currently not used to
compile MK but only its dependencies.

* Grok arm and implement arm flavours

* Fix arch conundrum and force min API to 9 where possible

* Make sure everything builds

* Add workaround for Android with API < 21

* build/android/cross: remove -Os from flags

As documented in https://github.com/measurement-kit/measurement-kit/issues/1051#issuecomment-269987809, the `-Os` flag leads to libevent binaries that do not work correctly. Let's remove this flag, rebuild and test again reporting the results into issue #1051.

* Pass -fno-integrated-as to armeabi

This fixes the segfault with armeabi documented in https://github.com/measurement-kit/measurement-kit/issues/1051#issuecomment-270107517.

Apparently, adding this flag to armeabi-v7a has no effect. That is, also with this flag applied, the emulator with API 10 still does not validate certificates.