system_resolver.hpp: do not warn on getaddrinfo failure (#1055)
This is too low level. For example, we can get an error here
because we're resolving AAAA and there is no associated record
but in fact there is a record for A and we can connect.

Issue noticed on measurement-kit/ios-example where this is the
last log message printed during each working test.

The general principle is that low level code should not warn or
emit log messages for failure unless it's unrecoverable.