web-connectivity: don't fail with non UTF-8 body
We cannot access the body through the entry, because, if it's not
UTF-8, it is represented as base64 and encoded as an object.

Instead, pass along the original response and use it.

Fixes crash noticed by @lorenzoPrimi on URL http://www.benedelman.org/.

While there, fix backend address used by web_connectivity command
to be able to actually test this fix.

Cc: @hellais.