var.hpp: fix as() to use dynamic cast
When developing a diff to improve NDT, I mistakenly wrote:

```C++
ErrorOr<std::vector<double>> ctimes = net::get_connect_times(err);
```

instead of:

```C++
ErrorOr<double> ctime = net::get_connect_time(err);
```

after `connect()`. Now, both functions internally use Var<T>::as()
to cast from a generic `ErrorContext` to a specific context.

The net effect of calling `get_connect_times` rather than the more
humble `get_connect_time` is that `ErrorContext` is downcasted to
`ConnectManyResult` rather than to `ConnectResult`. Since the cast
is not possible, one would have expected the cast to fail and the
resulting Var<T> to hold a nullptr. Hence the `ErrorOr` should not
be a value but rather an error. Right?

Wrong! What I obtained instead was a segfault.

It turns out that using a static-pointer cast in `as()` was a bad
idea, and that what I really would have wanted to have is a dynamic
cast instead. The dynamic cast, in fact, fails (as I expected) if
it is not possible to perform the cast.