Make sure network errors have a meaning
While investigating measurement-kit/ooniprobe-ios#61, I realized it
is now time to start improving the error codes.

This pull request starts implementing some ideas in #935 to make
at least errors emitted by the network layer much more clear.

Specifically, I've chosen to rely on the C++ library standard `errc`
enum (and to the corresponding names) for mapping errno.

To be honest, I dunno whether this is portable to Windows, but
the `std::errc` enum is described as `C++11` at:

    http://en.cppreference.com/w/cpp/error/errc

Moreover, libcxx <system_error> actually defines some of the errors
that we are interested to map in here:

    https://github.com/llvm-mirror/libcxx/blob/a7ae303de36dfa1e0a165c52cd073c51dc85e8cb/include/system_error#L262

So, this seems to me more solid that just peeking at `errno`.