connect_impl: fix treatment of IPv6 addresses
I believe, this diff alone should be enough to fix the issue
reported in measurement-kit/ooniprobe-ios#61.

I have noticed that also from my home (where there is no IPv6
connectivity) the code fails because it cannot even parse such
addresses. With this diff applied, instead, it fails because
of a `host_unreachable` error, which makes more sense to me (of
course, to see the real error, all of #1098 is required).