connect: start fixing IPv6 addresses
* connect_impl: fix treatment of IPv6 addresses

I believe, this diff alone should be enough to fix the issue
reported in measurement-kit/ooniprobe-ios#61.

I have noticed that also from my home (where there is no IPv6
connectivity) the code fails because it cannot even parse such
addresses. With this diff applied, instead, it fails because
of a `host_unreachable` error, which makes more sense to me (of
course, to see the real error, all of #1098 is required).

* connect: document scoped link-local IPv6 addrs bug

I was temped to fix this as part of this pull request but then
noticed it would have taken a bit of refactoring that instead I
would like to perform separately.

* connect: add warnings for each error case

We can (and will) do better by mapping errors, but, for now, let's
start by adding basic warning messages that end up in logs.