Make valgrind suppression file more general
We may from time to time refactor the regress test that fails
with Valgrind and Ubuntu 12.04, and we don't want the error to
resurface after each refactor just because the line number is
changed and so the rule does not match anymore. Just add a
wildcard such that we can refactor as much as we please.