http: fix EOF-implies-end-of-body logic (#1121)
We had logic to specially deal with EOF to signal end-of-body, however
the code was not resetting error after EOF resulting in inability to
process redirects where bodies were terminated by EOF.

This causes measurement-kit/ooniprobe-ios#79 reported by @willscott.

For the records, I also saw a failure for `hushmail.com` but I flagged
it as a transient network error. However, when also @willscott reported
the same error, it was clear this was something more than I thought.

While there, give a meaningful name to error 1000.

Also, since there is an open pull request refactoring HTTP, and
specifically I refer to measurement-kit/measurement-kit#1041, it
is wise to make sure this error is not re-introduced there.