Fix unknown_error_3009 bug
To this end, we need to add code for the case where the redirection is
towards a relative, rather than absolute, path.

As said in a comment, I thought this should not happen, but we've seen
cases in the wild where it does, so here we are with a patch.

While there, make use of mk::startswith() and mk::endswith() for
additional clarity.