http: treat EOF instead of response as error
As noted during #1174, it's odd that we return NoError() when
there is EOF and we've not determined whether body could be
terminated by EOF. That is actually a protocol violation more
than a case that we should treat as success.

Add test cases for both body terminated by EOF and EOF received
before anything else, so we're robust to future changes.