[ci skip] Completely remove all references
Now I understand the reason why the input vector was allocated
using `Var`. It was to keep it safe inside the closure, so that
references to its elements could be passed around.

I missed that and failed to understand why we were allocating
using `Var` and removed that to "improve" the code.

I think I can reinstate the original code quite easily.

But given that the input is small, we can just further simplify
code and always make copies.

It will be a little slower but we don't need to optimize this
code path that much.

I added a comment, so one is not tempted to optimize further by
reinstating the reference inside the loop.

The proper solution would probably be to use move semantic, but I
am not sure how to use move semantic together with range loops.

Anyway, sooner or later, I'll do a move semantic pass of the whole
tree, and in such case I'll also improve this code.

I am reluctant to play with move semantic now, since we're close
to release, and I don't want to screwup things.

For the records, running this code seems okay as Valgrind does not
complain (but it was not complaining even more, I just basically
refactored the code in a way that seems easier for me to maintain
in the future, with less tricky things to think at).