Implement Neubot's DASH test (#769)
Work originally done by Carmine D'Amico (@carminedamico) and later
reworked by @bassosimone. The PR with Carmine's work is #572.

Authorship of this commit is assigned to Carmine, as Simone mainly did
changes to update the code to make it work with current MK and refactoring
to simplify the code and adapt it to his tastes.