docker: use libc++ with clang (#1275)
* docker: use libc++ with clang

We are using very aggressive regular expressions in
web_connectivity. The problem is that the GNU C++
library has a very broken regex implementation that
uses stack recursion and segfaults when the expression
is too big. This happens for me on my development
`Mac mini` running Ubuntu zesty for `https://jezebel.com`
as input. I spotted this running pre-release checks.

So, I investigated ways to use libc++ also on Linux and
not only on macOS.  This diff modifies docker specs such
that, when we build with clang, we use libc++.

This opens up the possibility of packaging MK for debian
using libc++ rather than using GNU C++ library, because we
can see whether it is working or not.

Also updated the tutorial to explain how to build on a
generic Linux system using libc++.

See also: <https://stackoverflow.com/questions/36304204>.

* FIX test/http/request.cpp for hustmail

Basically: it seems that sometimes we are and sometimes we
are not redirected (judging from how regress fail).

Fix this by allowing for both cases.

Conflicts:
	test/http/request.cpp

* REPAIR test/http/request.cpp build

Apparently, Catch did not like the previous code: it did not compile.

This way, it should clearly compile.