fix(leak test): simply ignore exceptions in callbacks
The proposed approach was good but not good enough to make sure
that we were not leaking _all_.

I guess rethrowing the exception in the context of the callback was
not smart. It would perhaps have been better either to pass that
exception to a handler, or to register it in the poller, tear down
the loop and rethrow it _after_ the loop has gone down.