.travis.yml: add -Werror to CXXFLAGS and CFLAGS (#1316)
At least on the CI environment, I would like the code not to
produce any warning. In other scenarios, there _may_ be a few
warnings caused by porting software. For example, I do see
a few warnings about chars that cannot be negative on Android
on platforms that are negative. The code triggering such
warnings is not from us but from third parties.