fix(test/nettests): wait for background threads
We need to do so because now we allocate thread local storage
for SSL_CTX. Hence, if we don't wait for the thread to exit, we
have valgrind errors because such storage is not released.

The technique used in here is similar to the one that was also
employed for the orchestration.

Basically, we wait for the background tasks queue to be empty
in the ghetto way, i.e. busy looping.

Yet, the busy loop is short, because for the most time the code
will be blocked in run() and the busy loop is only to ensure that
we give enough time to background threads to finish.