fix(dash): disable HTTP caching
Some users have reported that DASH does not work for them when
run on mobile networks. And I could reproduce this behavior with
my Vodafone mobile connection.

Reading the logs, it seems to me that the main issue is a caching
issue where the cache deployed by the mobile operator kicks in.

My first attempt to try and resolve this issue is by adding an
explicity `Cache-Control: no-cache` header.

This may or may not fix the issue. Even if it does not fix the actual
issue, this is anyway something that we should add because it is
what the original implementation of Neubot does.