fix(net/libssl): better library initialization
Rather than constructing an object so to have library initialization
as a side effect, use explicit initialization function.

I was worried that a compiler could optimize away the object
construction because the object is not used, thus eliminating
also the side effect. Using a `void(void)` function seems to
me more safe, because it is obvious that you call it for a
purpose (the side effect), and so there is no grounds for a
compiler to optimize it away.